requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
boltons==24.0.0
cascade==0.1
catalogue==2.0.10
certifi==2024.6.2
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cryptography==43.0.0
cymem==2.0.8
Cython==3.0.10
deepface==0.0.92
defusedxml==0.8.0rc2
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==3.13.0
django-countries==7.5.1
django-formtools==2.5.1
django-money==3.5.3
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
# dlib @ file:///C:/Users/Asus/Downloads/Dlib_Windows_Python3.x-main/Dlib_Windows_Python3.x-main/dlib-19.24.1-cp311-cp311-win_amd64.whl
dnspython==2.6.1
docutils==0.21.2
face==22.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
gast==0.5.5
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
idna==3.7
imagesize==1.4.1
imaplib2==3.6
installer==0.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.4.1
Keras-Applications==1.0.8
keras-models==0.0.7
Keras-Preprocessing==1.1.2
langcodes==3.4.0
language_data==1.2.0
# libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
multidict==6.1.0
murmurhash==1.0.10
mysql-connector==2.2.9
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
preshed==3.0.9
progressbar==2.5
protobuf==4.25.3
py-moneyed==3.0
pyarrow==15.0.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.13.0
pymongo==4.8.0
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
reportlab==4.0.8
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.7.1
scipy==1.14.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.0.2
sphinx-autobuild==2024.4.16
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
spicy==0.16.0
sqlparse==0.5.0
srsly==2.4.8
starlette==0.38.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tf_keras==2.16.0
Theano==1.0.5
thinc==8.2.5
tqdm==4.66.4
twilio==9.3.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
vaderSentiment==3.3.2
virtualenv==20.0.31
wasabi==1.1.3
watchfiles==0.23.0
weasel==0.4.1
websockets==13.0
Werkzeug==3.0.3
whitenoise==6.7.0
wrapt==1.16.0
yarl==1.13.1
